refactor code
split js files into separate files
